requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==4.2.7
django-filter==23.3
django-rest-framework==0.1.0
djangorestframework==3.14.0
factory-boy==3.3.0
Faker==20.0.3
inflection==0.5.1
iniconfig==2.0.0
packaging==23.2
Pillow==10.1.0
pluggy==1.3.0
pytest==7.4.3
pytest-django==4.7.0
pytest-factoryboy==2.6.0
pytest-mock==3.12.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
